Reuse maxSlope computation
